+++ src/modules/cache/mod.ts
@@ -7,6 +7,16 @@
  * \_____/ \__,_||_| |_||_|/_/\_\
  */
 
+import type { RedisClientType } from 'redis'
+
 export { ZanixCacheCoreProvider } from './providers/mod.ts'
 export { ZanixQLRUConnector } from './providers/qlru/connector.ts'
 export { ZanixRedisConnector } from './providers/redis/connector/mod.ts'
+
+// types
+export type {
+  /**
+   * Redis client type
+   */
+  RedisClientType,
+}